refactor: move ConnectionManager into its own module
