+++ ably/types/connectionstate.py
@@ -1,0 +1,36 @@
+from enum import Enum
+from dataclasses import dataclass
+from typing import Optional
+
+from ably.util.exceptions import AblyException
+
+
+class ConnectionState(str, Enum):
+    INITIALIZED = 'initialized'
+    CONNECTING = 'connecting'
+    CONNECTED = 'connected'
+    DISCONNECTED = 'disconnected'
+    CLOSING = 'closing'
+    CLOSED = 'closed'
+    FAILED = 'failed'
+    SUSPENDED = 'suspended'
+
+
+class ConnectionEvent(str, Enum):
+    INITIALIZED = 'initialized'
+    CONNECTING = 'connecting'
+    CONNECTED = 'connected'
+    DISCONNECTED = 'disconnected'
+    CLOSING = 'closing'
+    CLOSED = 'closed'
+    FAILED = 'failed'
+    SUSPENDED = 'suspended'
+    UPDATE = 'update'
+
+
+@dataclass
+class ConnectionStateChange:
+    previous: ConnectionState
+    current: ConnectionState
+    event: ConnectionEvent
+    reason: Optional[AblyException] = None  # RTN4f
